hive_config_manager/requirements/requirements.txt
# requirements/requirements.txt

# Core dependencies
packaging>=23.2
PyYAML>=6.0.1
prompt-toolkit>=3.0.43
inquirer>=3.1.4
pytz>=2023.3.post1

# Command line interface
rich>=13.7.0
typer>=0.9.0

# File operations and system utilities
filelock>=3.13.1
pathlib>=1.0.1

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
